pyproject.toml
[tool.poetry]
name = "relatorio_de_gastos"
version = "0.1.0"
description = "Um web app para monitorar os gastos diários em um dashboard."
authors = ["Wagno Leão Sergio <wagnoleao@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6.1,<4.0"
numpy = "^1.19.4"
pandas = "^1.1.5"
matplotlib = "^3.3.3"
plotly = "^4.14.0"
streamlit = "^0.72.0"
xlrd = "^1.2.0"
seaborn = "^0.11.0"
mongoengine = "^0.21.0"
dnspython = "^2.0.0"
python-dotenv = "^0.15.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
